enhancement: add a notice filter to warn only if feed_contact_email or url are both unset
