+++ core/src/main/java/org/mobilitydata/gtfsvalidator/NoticeFilter/NoticeFilter.java
@@ -1,12 +1,13 @@
 package org.mobilitydata.gtfsvalidator.NoticeFilter;
 
-import org.mobilitydata.gtfsvalidator.notice.NoticeContainer;
 import org.mobilitydata.gtfsvalidator.notice.ValidationNotice;
-import org.mobilitydata.gtfsvalidator.parsing.RowParser;
 
 import java.util.List;
 
+/***
+ * For apply some rules to filter validation notices
+ */
 public interface NoticeFilter {
-    public abstract void filter(NoticeContainer noticeContainer);
+    public abstract void filter(List<ValidationNotice> validationNotices);
 
 }
