+++ core/src/main/java/org/mobilitydata/gtfsvalidator/table/AnyTableLoader.java
@@ -11,7 +11,6 @@ import java.util.ServiceLoader;
 import java.util.stream.Collectors;
 import javax.annotation.Nullable;
 
-import org.mobilitydata.gtfsvalidator.NoticeFilter.NoticeFilter;
 import org.mobilitydata.gtfsvalidator.notice.CsvParsingFailedNotice;
 import org.mobilitydata.gtfsvalidator.notice.EmptyFileNotice;
 import org.mobilitydata.gtfsvalidator.notice.MissingRecommendedFileNotice;
@@ -111,10 +110,7 @@ public final class AnyTableLoader {
 
         noticeContainer.addAll(rowNotices);
 
-        ServiceLoader<NoticeFilter> noticeFilterServiceLoader = ServiceLoader.load(NoticeFilter.class);
-        for(NoticeFilter filter:noticeFilterServiceLoader){
-          filter.filter(noticeContainer);
-        }
+
       }
     } catch (TextParsingException e) {
       noticeContainer.addValidationNotice(new CsvParsingFailedNotice(gtfsFilename, e));
