+++ main/build.gradle
@@ -35,6 +35,8 @@ dependencies {
     compileOnly project(':processor')
     implementation project(':model')
     implementation project(':core')
+    annotationProcessor 'com.google.auto.service:auto-service:1.0-rc4'
+    compileOnly 'com.google.auto.service:auto-service:1.0-rc4'
     annotationProcessor 'com.google.auto.value:auto-value:1.7.4'
     compileOnly 'com.google.auto.value:auto-value-annotations:1.7.4'
     implementation 'org.apache.httpcomponents.client5:httpclient5:5.0.3'
